Make ConlluReader child of GCReader
